requirements.txt
fastapi
uvicorn
python-dotenv
firebase_admin
requests
Pillow
pydantic
google-generativeai
pywhatkit
opencv-python
pytesseract
